+++ src/main/java/radon/jujutsu_kaisen/ability/curse_manipulation/util/CurseManipulationUtil.java
@@ -43,11 +43,6 @@ public class CurseManipulationUtil {
         Optional<GameProfile> profile = curse.profile();
         profile.ifPresent(gameProfile -> ((AbsorbedPlayerEntity) entity).setPlayer(gameProfile));
 
-        Vec3 pos = owner.position()
-                .subtract(RotationUtil.calculateViewVector(0.0F, owner.getYRot())
-                        .multiply(entity.getBbWidth() / 2.0F, 0.0D, entity.getBbWidth() / 2.0F));
-        entity.moveTo(pos.x, pos.y, pos.z, owner.getYRot(), owner.getXRot());
-
         return entity;
     }
 
@@ -107,6 +102,11 @@ public class CurseManipulationUtil {
         CursedSpirit entity = createCurse(owner, curse);
 
         if (entity == null) return null;
+
+        Vec3 pos = owner.position()
+                .subtract(RotationUtil.calculateViewVector(0.0F, owner.getYRot())
+                        .multiply(entity.getBbWidth() / 2.0F, 0.0D, entity.getBbWidth() / 2.0F));
+        entity.moveTo(pos.x, pos.y, pos.z, owner.getYRot(), owner.getXRot());
 
         owner.level().addFreshEntity(entity);